[KAN-27] refactor: 필터 버튼 컴포넌트화
